chore: Phase 1
- Initialize shadcn/ui with 10 components (button, badge, dialog, input,
  label, table, select, tabs, separator, tooltip)
- Add husky + lint-staged pre-commit hooks (eslint, prettier on staged files)
- Fix Rust compilation errors: add missing use tauri::Manager imports
- Fix tsconfig.node.json: add composite + emitDeclarationOnly for project refs
- Generate app icons via npx tauri icon
- Update .gitignore with comprehensive ignore patterns
- Fix all clippy warnings: rustfmt, unused vars/imports, dead code, Box<PageRow>
- Fix eslint config: install typescript-eslint, update lint scripts for flat config
- Add frontend unit tests for utility functions (15 tests)
- Add jsdom dependency for vitest environment
- Add passWithNoTests to vitest config
- Run prettier on all source files for consistent formatting
- Add continue-on-error to CI integration test step (no tests yet)
- Upgrade cn() to use clsx + tailwind-merge for Tailwind conflict resolution

+++ src-tauri/benches/frontier_ops.rs
@@ -1,5 +1,5 @@
-use criterion::{black_box, criterion_group, criterion_main, Criterion};
-use oxide_seo_lib::crawler::frontier::{hash_url, normalize_url, FrontierEntry, UrlFrontier};
+use criterion::{Criterion, black_box, criterion_group, criterion_main};
+use oxide_seo_lib::crawler::frontier::{FrontierEntry, UrlFrontier, hash_url, normalize_url};
 
 fn bench_normalize_url(c: &mut Criterion) {
     c.bench_function("normalize_url", |b| {
@@ -30,7 +30,7 @@ fn bench_frontier_push_pop(c: &mut Criterion) {
                     url,
                     url_hash: hash,
                     depth: (i % 5) as u32,
-                    priority: 100 - (i % 5) as i32,
+                    priority: 100 - (i % 5),
                     source_page_id: None,
                 });
             }
@@ -41,5 +41,10 @@ fn bench_frontier_push_pop(c: &mut Criterion) {
     });
 }
 
-criterion_group!(benches, bench_normalize_url, bench_hash_url, bench_frontier_push_pop);
+criterion_group!(
+    benches,
+    bench_normalize_url,
+    bench_hash_url,
+    bench_frontier_push_pop
+);
 criterion_main!(benches);
